[MOV] html_builder, website: move generic save_snippet
This commit moves all the generic parts of the save snippet process to
html_builder. These are necessary for the addition of the process to
mass_mailing, as some parts of it required website to work.

Now, the SaveSnippetPlugin can be used in any context as the relevant website
parts of the code were moved back to website.
We also introduced a new helper to patch a specific SnippetModel Object, as the
original patch was applied to the prototype. Thereby affecting models unrelated
to website, e.g. `mass_mailing`.
Now, only the object linked to website will be impacted by the website patch.

task-5189391
