+++ addons/html_builder/models/__init__.py
@@ -1,1 +1,2 @@
 from . import ir_qweb
+from . import ir_ui_view
