+++ addons/html_builder/static/src/snippets/snippet_service.js
@@ -1,12 +1,13 @@
 import { ConfirmationDialog } from "@web/core/confirmation_dialog/confirmation_dialog";
 import { _t } from "@web/core/l10n/translation";
-import { uniqueId } from "@web/core/utils/functions";
+import { memoize, uniqueId } from "@web/core/utils/functions";
 import { Reactive } from "@web/core/utils/reactive";
 import { AddSnippetDialog } from "./add_snippet_dialog";
 import { registry } from "@web/core/registry";
 import { user } from "@web/core/user";
 import { markup, useState } from "@odoo/owl";
 import { useService } from "@web/core/utils/hooks";
+import { patch } from "@web/core/utils/patch";
 
 export class SnippetModel extends Reactive {
     constructor(services, { snippetsName, context }) {
@@ -256,11 +257,122 @@ export class SnippetModel extends Reactive {
     }
 
     async deleteCustomSnippet(snippet) {
-        throw new Error("Method not implemented");
+        return new Promise((resolve) => {
+            const message = _t(
+                "Are you sure you want to delete %s from your list of custom blocks?\nDeleting it will remove it for all users.",
+                snippet.title
+            );
+            this.dialog.add(
+                ConfirmationDialog,
+                {
+                    body: message,
+                    confirm: async () => {
+                        const isInnerContent =
+                            this.snippetsByCategory.snippet_custom_content.includes(snippet);
+                        const snippetCustom = isInnerContent
+                            ? this.snippetsByCategory.snippet_custom_content
+                            : this.snippetsByCategory.snippet_custom;
+                        const index = snippetCustom.findIndex((s) => s.id === snippet.id);
+                        if (index > -1) {
+                            snippetCustom.splice(index, 1);
+                        }
+                        await this.orm.call("ir.ui.view", "delete_snippet", [], {
+                            view_id: snippet.viewId,
+                            template_key: this.snippetsName,
+                        });
+                    },
+                    cancel: () => {},
+                    confirmLabel: _t("Delete Block"),
+                    cancelLabel: _t("Keep it"),
+                    title: _t("Delete this block?"),
+                },
+                {
+                    onClose: resolve,
+                }
+            );
+        });
     }
 
+    /**
+     * @override
+     */
     async renameCustomSnippet(snippet, newName) {
-        throw new Error("Method not implemented");
+        if (newName === snippet.title) {
+            return;
+        }
+        snippet.title = newName;
+        for (const snippetEl of this.snippetsDocument.body.querySelectorAll(
+            `snippets#snippet_custom > [data-oe-snippet-key=${snippet.key}]`
+        )) {
+            snippetEl.setAttribute("name", newName);
+            snippetEl.firstElementChild.dataset["name"] = newName;
+        }
+        await this.orm.call("ir.ui.view", "rename_snippet", [], {
+            name: newName,
+            view_id: snippet.viewId,
+            template_key: this.snippetsName,
+        });
+    }
+
+    getContext() {
+        return { ...this.context };
+    }
+
+    cleanSnippetForSave(snippetCopyEl, cleanForSaveHandlers) {
+        cleanForSaveHandlers.forEach((handler) => handler({ root: snippetCopyEl }));
+    }
+
+    /**
+     * Saves the given snippet as a custom one and reloads all the snippets
+     * to have access to it directly.
+     *
+     * @param {HTMLElement} snippetEl the snippet we want to save
+     * @param {Array<Function>} cleanForSaveHandlers all the handlers of the
+     *     `clean_for_save_handlers` resources
+     * @param {Function} wrapWithSaveSnippetHandlers a function that processes the snippet
+     * before and/or after the cloning. E.g. stopping the interactions before
+     * cloning and restarting them after cloning.
+     * @returns
+     */
+    async saveSnippet(
+        snippetEl,
+        cleanForSaveHandlers,
+        wrapWithSaveSnippetHandlers = (_, callback) => callback()
+    ) {
+        const isButton = snippetEl.matches("a.btn");
+        const snippetKey = isButton ? "s_button" : snippetEl.dataset.snippet;
+        const thumbnailURL = this.getSnippetThumbnailURL(snippetKey);
+
+        const snippetCopyEl = await wrapWithSaveSnippetHandlers(snippetEl, () =>
+            snippetEl.cloneNode(true)
+        );
+        // "CleanForSave" the snippet copy
+        this.cleanSnippetForSave(snippetCopyEl, cleanForSaveHandlers);
+
+        const defaultSnippetName = isButton
+            ? _t("Custom Button")
+            : _t("Custom %s", snippetEl.dataset.name);
+        snippetCopyEl.classList.add("s_custom_snippet");
+        delete snippetCopyEl.dataset.name;
+        if (isButton) {
+            snippetCopyEl.classList.remove("mb-2");
+            snippetCopyEl.classList.add("o_snippet_drop_in_only", "s_custom_button");
+        }
+
+        const context = this.getContext(snippetEl);
+
+        const savedName = await this.orm.call("ir.ui.view", "save_snippet", [], {
+            name: defaultSnippetName,
+            arch: snippetCopyEl.outerHTML,
+            template_key: this.snippetsName,
+            snippet_key: snippetKey,
+            thumbnail_url: thumbnailURL,
+            context,
+        });
+
+        // Reload the snippets so the sidebar is up to date.
+        await this.reload();
+        return savedName;
     }
 
     setSnippetName(snippetsDocument) {
@@ -307,26 +419,6 @@ export class SnippetModel extends Reactive {
     }
 
     /**
-     * Saves the given snippet as a custom one and reloads all the snippets
-     * to have access to it directly.
-     *
-     * @param {HTMLElement} snippetEl the snippet we want to save
-     * @param {Array<Function>} cleanForSaveHandlers all the hanlders of the
-     *     `clean_for_save_handlers` resources
-     * @param {Function} wrapWithSaveSnippetHandlers a function that processes the snippet
-     * before and/or after the cloning. E.g. stopping the interactions before
-     * cloning and restarting them after cloning.
-     * @returns
-     */
-    saveSnippet(
-        snippetEl,
-        cleanForSaveHandlers,
-        wrapWithSaveSnippetHandlers = (_, callback) => callback()
-    ) {
-        throw new Error("Method not implemented");
-    }
-
-    /**
      * Gets the label of the snippet.
      *
      * @param {HTMLElement} snippetEl
@@ -361,8 +453,12 @@ export const snippetService = {
             );
             return snippetModelsMap.get(snippetsName);
         };
+        // patchSnippetModel only supports a single patch per snippetModel, to refactor if more are needed.
+        const patchSnippetModel = memoize((snippetsName, snippetModelPatch) => {
+            patch(getSnippetModel(snippetsName), snippetModelPatch);
+        });
 
-        return { getSnippetModel };
+        return { getSnippetModel, patchSnippetModel };
     },
 };
